:recycle: Refactor and format

+++ script/DexTwo.s.sol
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.13;
 
-import "src/levels/DexTwo/DexTwo.sol"; 
+import "src/levels/DexTwo/DexTwo.sol";
 import "forge-std/Script.sol";
 
 contract DexTwoAttackScript is Script {
@@ -9,7 +9,7 @@ contract DexTwoAttackScript is Script {
 
     function run() public {
         DexTwo target = DexTwo(0x0000000000000000000000000000000000000000); // Insert instance target address
-                             
+
         vm.startBroadcast();
 
         // Deploy MaliciousTokenContract instance
@@ -24,11 +24,11 @@ contract DexTwoAttackScript is Script {
 }
 
 contract MaliciousTokenContract {
-    function balanceOf(address) public pure returns (uint){
+    function balanceOf(address) public pure returns (uint256) {
         return 100;
     }
 
-    function transferFrom(address, address, uint) public pure returns(bool) {
+    function transferFrom(address, address, uint256) public pure returns (bool) {
         return true;
     }
 }
